Recode to `RUF043`
