+++ crates/ruff_linter/resources/test/fixtures/ruff/RUF043.py
