+++ crates/ruff_linter/src/rules/ruff/rules/mod.rs
@@ -23,6 +23,7 @@ pub(crate) use never_union::*;
 pub(crate) use none_not_at_end_of_union::*;
 pub(crate) use parenthesize_chained_operators::*;
 pub(crate) use post_init_default::*;
+pub(crate) use pytest_raises_ambiguous_pattern::*;
 pub(crate) use quadratic_list_summation::*;
 pub(crate) use redirected_noqa::*;
 pub(crate) use redundant_bool_literal::*;
@@ -71,6 +72,7 @@ mod never_union;
 mod none_not_at_end_of_union;
 mod parenthesize_chained_operators;
 mod post_init_default;
+mod pytest_raises_ambiguous_pattern;
 mod quadratic_list_summation;
 mod redirected_noqa;
 mod redundant_bool_literal;
