+++ ruff.schema.json
@@ -3676,9 +3676,6 @@
         "PT025",
         "PT026",
         "PT027",
-        "PT2",
-        "PT20",
-        "PT201",
         "PTH",
         "PTH1",
         "PTH10",
@@ -3851,6 +3848,7 @@
         "RUF04",
         "RUF040",
         "RUF041",
+        "RUF043",
         "RUF046",
         "RUF048",
         "RUF05",
